Rename players to aimmo (#652)
* Rename players to aimmo

* players -> aimmo
