+++ aimmo/_version.py
@@ -38,7 +38,7 @@ def get_config():
     cfg.style = "pep440"
     cfg.tag_prefix = ""
     cfg.parentdir_prefix = "aimmo-"
-    cfg.versionfile_source = "players/_version.py"
+    cfg.versionfile_source = "aimmo/_version.py"
     cfg.verbose = False
     return cfg
