+++ aimmo/avatar_examples/dumb_avatar.py
