+++ aimmo/avatar_examples/health_seeker_avatar.py
