+++ aimmo/management/commands/delete_generated_players.py
