+++ aimmo/tests.py
@@ -6,7 +6,7 @@ from django.contrib.auth.models import AnonymousUser, User
 from django.core.urlresolvers import reverse
 from django.test import Client, TestCase
 
-from players import models, app_settings
+from aimmo import models, app_settings
 
 app_settings.GAME_SERVER_URL_FUNCTION = lambda game_id: ('base %s' % game_id, 'path %s' % game_id)
 app_settings.GAME_SERVER_PORT_FUNCTION = lambda game_id: 0
