+++ aimmo_runner/setup.cfg
@@ -7,7 +7,7 @@ max-line-length = 160
 [versioneer]
 VCS = git
 style = pep440-pre
-versionfile_source = ../players/_version.py
-versionfile_build = ../players/_version.py
+versionfile_source = ../aimmo/_version.py
+versionfile_build = ../aimmo/_version.py
 tag_prefix =
 parentdir_prefix = aimmo-
