+++ game_frontend/src/redux/api/constants.js
@@ -1,1 +1,1 @@
-export const API_PATH = '/players/api/'
+export const API_PATH = '/aimmo/api/'
