+++ setup.cfg
@@ -2,11 +2,11 @@
 universal = 1
 
 [coverage:run]
-source = players, aimmo-game, aimmo-game-worker, aimmo-game-creator
+source = aimmo, aimmo-game, aimmo-game-worker, aimmo-game-creator
 
 [coverage:report]
 omit =
-    players/_version.py
+    aimmo/_version.py
     *test-*
     */tests/*
     */python?.?/*
@@ -19,7 +19,7 @@ max-line-length = 160
 [versioneer]
 VCS = git
 style = pep440-pre
-versionfile_source = players/_version.py
-versionfile_build = players/_version.py
+versionfile_source = aimmo/_version.py
+versionfile_build = aimmo/_version.py
 tag_prefix =
 parentdir_prefix = aimmo-
